pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["spawlov <107248409+spawlov@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = {extras = ["standart"], version = "^0.34.2"}
pydantic = {extras = ["email"], version = "^2.11.5"}
pydantic-settings = "^2.9.1"
alembic = "^1.16.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
aiosqlite = "^0.21.0"
argon2-cffi = "^23.1.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
skip_gitignore = true
order_by_type = false

[tool.flake8]
ignore = [
    'D100',  # Missing docstring in public module
    'D101',  # Missing docstring in public class
    'D102',  # Missing docstring in public method
    'D103',  # Missing docstring in public function
    'D104',  # Missing docstring in public package
    'D105',  # Missing docstring in magic method
    'D106',  # Missing docstring in public nested class
    'D107',  # Missing docstring in __init__
    'S101',
    'W503',  # Line break occurred before a binary operator, conflicts with W504:
             # https://github.com/mila-iqia/cookiecutter-pyml/issues/10
]
max-line-length = 120
max-complexity = 7
exclude = "migrations"  # exclude autogenerated Django migrations
count = true
import-order-style = "edited"
